Change default container name to "container.sif"
Set the default "output_file" name of the Apptainer action to
"container.sif" and also set this as default for all the other actions.

This way, the container will by default always have the same name,
independent of the name in the URI that is pulled.  Thus one does not
need to make further changes on other actions when using a different
container.