Place colcon build dir outside of working directory
In the workflow `colcon build` is called inside the package directory.
This means that also the `install` and `build` directories are created
there.  The build directory contains some C++ files that are generated
by CMake, apparently to identify the compiler.  If kept inside the
package directory, those files are also found by the documentation build
and automatically added to the C++ API docs, which is not desired.

As an easy fix, simply locate the build directory in /tmp (it is anyway
not needed anymore after the build is finished).